Fixes #8820 improvement the deleted contentlet message

+++ src/com/dotmarketing/portlets/cmsmaintenance/ajax/CMSMaintenanceAjax.java
@@ -7,6 +7,7 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.sql.SQLException;
+import java.text.MessageFormat;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
@@ -21,6 +22,7 @@ import java.util.zip.ZipOutputStream;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 
+import com.dotmarketing.portlets.contentlet.business.ImpossibleDeleteArchivedContentletException;
 import com.dotmarketing.portlets.rules.util.RulesImportExportUtil;
 import org.quartz.JobExecutionContext;
 
@@ -199,14 +201,11 @@ public class CMSMaintenanceAjax {
 
 
 
-	public String[] deleteContentletsFromIdList(String List, String userId) throws PortalException, SystemException, DotDataException,DotSecurityException {
-
+	public String deleteContentletsFromIdList(String List, String userId) throws PortalException, SystemException, DotDataException,DotSecurityException {
+		List<String> conditionletWithErrors = new ArrayList<>();
 		ContentletAPI conAPI = APILocator.getContentletAPI();
 		String[] inodes = List.split(",");
 		Integer contdeleted = 0;
-		String contnotfound = "";
-		String conthasreqrel = "";
-		String conthasnoperm = "";
 
 		User user = UserLocalManagerUtil.getUserById(userId);
 		for (int i = 0; i < inodes.length; i++) {
@@ -221,14 +220,44 @@ public class CMSMaintenanceAjax {
 			}
 		}
 
-		for (Contentlet contentlet : contentlets) {
-			conAPI.delete(contentlet, APILocator.getUserAPI().getSystemUser(), true, true);
-			contdeleted++;
+		if (!contentlets.isEmpty()) {
+			for (Contentlet contentlet : contentlets) {
+				try {
+					conAPI.delete(contentlet, APILocator.getUserAPI().getSystemUser(), true, true);
+					contdeleted++;
+				} catch (ImpossibleDeleteArchivedContentletException e) {
+					conditionletWithErrors.add(contentlet.getIdentifier());
+				}
+			}
+
+			return getDeleteContentletMessage(user, conditionletWithErrors, contdeleted);
+		}else{
+			return LanguageUtil.get(user, "message.contentlet.delete.error.dontExists");
 		}
 
-		String[] results = { contdeleted.toString(), contnotfound, conthasreqrel,conthasnoperm };
 
-		return results;
+	}
+
+	private String getDeleteContentletMessage(User user, List<String> conditionletWithErrors, Integer contdeleted) throws LanguageException {
+		String result = null;
+
+		if (conditionletWithErrors.isEmpty()){
+			result = LanguageUtil.get(user,"contentlets-were-succesfully-deleted");
+		}else{
+			String errorMessage = LanguageUtil.get(user,"message.contentlet.delete.error.archived");
+			String conditionletIdentifier = conditionletWithErrors.toString().replace("[", "").replace("]", "")
+					.replace(", ", ",");
+			errorMessage = MessageFormat.format(errorMessage, conditionletIdentifier);
+
+			if (contdeleted > 0){
+				result = LanguageUtil.get(user, "message.contentlet.delete.success.withError");
+				result = MessageFormat.format(result, errorMessage);
+			}else{
+				result = errorMessage;
+			}
+		}
+
+		return result;
 	}
 
 	public String deletePushedAssets() throws PortalException, SystemException, DotDataException,DotSecurityException {
